Equations no longer mark their auxiliary varaibles.
For input equations that are of type Equality, no aux variable is
needed.
For equations of types GE or LE, the preprocessor adds the auxiliary
varibale and adjusts the bounds.

For equations that are adde during the run (e.g., splits), the tableau
always allocates an aux variable, even for equalities - but this again
is transparent to the equation itself and to the PL constraints.