Refactor _phaseStatus from individual classes into PiecewiseLinearConstraint (#404)
* Add phase status member to PiecewiseLinearConstraint class

All PiecewiseLinearConstraints have a phase/case, either a fixed number of them (ReLU, Sign, Abs) or a flexible number of them (Max, Disjunction). The first group cares about named elements of the enumeration, while the second group cares about their numerical value, and a fixed size enumeration caters to both. Regardless of the subclass, the role of these fields is the same and becomes further relevant in context-dependent implementation for detection of implications, contradictions, etc.

The idea is to generalize this field to avoid repetition of identical code in child classes, and level the implementation so that a single intermediate child class implement the context-dependent behavior, while retaining existing functionality for preprocessing purposes. (The intermediate class can be merged with the parent class at a later date).

* Remove local phase status members in derivations of PiecewiseLinearConstraint

Remove calls to initialize _phaseStatus in constructors, this is now done by
constructor of PiecewiseLinearConstraint.

Refactor all use of local PhaseStatus enumerations, to use the joint
PhaseStatus enumeration.

Remove local PhaseStatus enumerations.

* Fix empty line

* Fix typo